+++ changelog.md
@@ -2,3 +2,5 @@
 - added back dispensers behaviors
 - added ryoamic lights compat
 - fixed candle holder in hand flame
+- improved sign screen renderer to render block model
+- fixed wall sign model being 1 pixel too tall